requirements.txt
python-poppler
xlsxwriter~=3.2.0
python-magic~=0.4.27
tkinterdnd2~=0.3.0
pandas~=2.2.2